refactor(fp): extract collection drag controller into useFPCollectionDnd hook
